requirements.txt
psutil
pandas
Pillow
matplotlib
#tkinter
